Move runtimeResult and Add Contract Invoke Test
