+++ yggdrash-core/src/main/java/io/yggdrash/core/blockchain/BlockChain.java
@@ -21,7 +21,7 @@ import io.yggdrash.core.contract.Contract;
 import io.yggdrash.core.exception.FailedOperationException;
 import io.yggdrash.core.exception.InvalidSignatureException;
 import io.yggdrash.core.exception.NotValidateException;
-import io.yggdrash.core.runtime.BlockRuntimeResult;
+import io.yggdrash.core.runtime.result.BlockRuntimeResult;
 import io.yggdrash.core.runtime.Runtime;
 import io.yggdrash.core.store.BlockStore;
 import io.yggdrash.core.store.MetaStore;
